renamed SignSufficient Crypto to sign specs
