+++ android/src/main/java/io/parity/signer/screens/settings/networks/signspecs/SignSpecsFull.kt
@@ -1,13 +1,13 @@
-package io.parity.signer.screens.settings.networks.sufficientcrypto
+package io.parity.signer.screens.settings.networks.signspecs
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import io.parity.signer.screens.settings.networks.sufficientcrypto.view.SufficientCryptoReadyBottomSheet
+import io.parity.signer.screens.settings.networks.signspecs.view.SufficientCryptoReadyBottomSheet
 import io.parity.signer.bottomsheets.password.EnterPassword
 import io.parity.signer.domain.Callback
-import io.parity.signer.screens.settings.networks.sufficientcrypto.view.SignSufficientCryptoScreen
+import io.parity.signer.screens.settings.networks.signspecs.view.SignSufficientCryptoScreen
 import io.parity.signer.ui.BottomSheetWrapperRoot
 import io.parity.signer.uniffi.MSignSufficientCrypto
 
@@ -18,11 +18,11 @@ import io.parity.signer.uniffi.MSignSufficientCrypto
 // end of action here calling go forward and it's in navstate.rs:427
 
 @Composable
-fun SignSufficientCryptoFull(
+fun SignSpecsFull(
 	sc: MSignSufficientCrypto,
 	onBack: Callback
 ) {
-	val vm: SignSufficientCryptoViewModel = viewModel()
+	val vm: SignSpecsViewModel = viewModel()
 
 	val passwordState = vm.password.collectAsStateWithLifecycle()
 	val signatureState = vm.signature.collectAsStateWithLifecycle()
