+++ android/src/main/java/io/parity/signer/screens/settings/networks/signspecs/view/SignSufficientCryptoScreen.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.screens.settings.networks.sufficientcrypto.view
+package io.parity.signer.screens.settings.networks.signspecs.view
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
