+++ android/src/main/java/io/parity/signer/screens/settings/networks/signspecs/view/SufficientCryptoReadyViewModel.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.screens.settings.networks.sufficientcrypto.view
+package io.parity.signer.screens.settings.networks.signspecs.view
 
 import io.parity.signer.domain.backend.mapError
 import io.parity.signer.dependencygraph.ServiceLocator
